created components for header, todo, footer and errors
