+++ src/components/ErrorNotification/ErrorNotification.tsx
@@ -1,0 +1,35 @@
+import cn from 'classnames';
+
+type Props = {
+  error: string;
+  setError: React.Dispatch<React.SetStateAction<string>>;
+};
+
+export const ErrorNotification: React.FC<Props> = ({ error, setError }) => {
+  return (
+    <div
+      data-cy="ErrorNotification"
+      className={cn('notification is-danger is-light has-text-weight-normal', {
+        hidden: !error,
+      })}
+    >
+      <button
+        data-cy="HideErrorButton"
+        type="button"
+        className="delete"
+        onClick={() => setError('')}
+      />
+      {/* show only one message at a time
+        Unable to load todos
+        <br />
+        Title should not be empty
+        <br />
+        Unable to add a todo
+        <br />
+        Unable to delete a todo
+        <br />
+        Unable to update a todo*/}
+      {error}
+    </div>
+  );
+};
