+++ src/components/Footer/Footer.tsx
@@ -1,0 +1,74 @@
+import { Filter } from '../../types/Filter';
+import { Todo } from '../../types/Todo';
+import cn from 'classnames';
+
+type Props = {
+  filter: Filter;
+  activeCount: number;
+  setFilter: React.Dispatch<React.SetStateAction<Filter>>;
+  visibleTodos: Todo[];
+  setTodos: React.Dispatch<React.SetStateAction<Todo[]>>;
+};
+
+export const Footer: React.FC<Props> = ({
+  filter,
+  activeCount,
+  setFilter,
+  visibleTodos,
+  setTodos,
+}) => {
+  return (
+    <footer className="todoapp__footer" data-cy="Footer">
+      <span className="todo-count" data-cy="TodosCounter">
+        {activeCount} items left
+      </span>
+
+      {/* Active link should have the 'selected' class */}
+      <nav className="filter" data-cy="Filter">
+        <a
+          href="#/"
+          className={cn('filter__link', {
+            selected: filter === 'all',
+          })}
+          data-cy="FilterLinkAll"
+          onClick={() => setFilter('all')}
+        >
+          All
+        </a>
+
+        <a
+          href="#/active"
+          className={cn('filter__link', {
+            selected: filter === 'active',
+          })}
+          data-cy="FilterLinkActive"
+          onClick={() => setFilter('active')}
+        >
+          Active
+        </a>
+
+        <a
+          href="#/completed"
+          className={cn('filter__link', {
+            selected: filter === 'completed',
+          })}
+          data-cy="FilterLinkCompleted"
+          onClick={() => setFilter('completed')}
+        >
+          Completed
+        </a>
+      </nav>
+
+      {/* this button should be disabled if there are no completed todos */}
+      <button
+        type="button"
+        className="todoapp__clear-completed"
+        data-cy="ClearCompletedButton"
+        disabled={visibleTodos.every(todo => !todo.completed)}
+        onClick={() => setTodos(todos => todos.filter(todo => !todo.completed))}
+      >
+        Clear completed
+      </button>
+    </footer>
+  );
+};
