+++ src/components/ToDo/ToDo.tsx
@@ -1,0 +1,44 @@
+/* eslint-disable jsx-a11y/label-has-associated-control */
+import cn from 'classnames';
+import { Todo } from '../../types/Todo';
+
+type Props = {
+  handleActive: (value: number) => void;
+  todo: Todo;
+};
+
+export const ToDo: React.FC<Props> = ({ handleActive, todo }) => {
+  return (
+    <div
+      data-cy="Todo"
+      className={cn('todo', {
+        completed: todo.completed,
+      })}
+    >
+      <label className="todo__status-label" htmlFor={`${todo.id}`}>
+        <input
+          data-cy="TodoStatus"
+          type="checkbox"
+          id={`${todo.id}`}
+          className="todo__status"
+          checked={todo.completed}
+          onChange={() => handleActive(todo.id as number)}
+        />
+      </label>
+
+      <span data-cy="TodoTitle" className="todo__title">
+        {todo.title}
+      </span>
+
+      {/* Remove button appears only on hover */}
+      <button type="button" className="todo__remove" data-cy="TodoDelete">
+        ×
+      </button>
+      {/* overlay will cover the todo while it is being deleted or updated */}
+      <div data-cy="TodoLoader" className="modal overlay">
+        <div className="modal-background has-background-white-ter" />
+        <div className="loader" />
+      </div>
+    </div>
+  );
+};
