+++ src/types/Filter.ts
@@ -1,0 +1,1 @@
+export type Filter = 'all' | 'active' | 'completed';
