+++ src/types/Todo.ts
@@ -1,5 +1,5 @@
 export interface Todo {
-  id?: number;
+  id: number;
   userId: number;
   title: string;
   completed: boolean;
